src/test/java/com/niuniu/shinetea/service/impl/CategoryServiceImplTest.java
package com.niuniu.shinetea.service.impl;

import com.niuniu.shinetea.dataobject.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
public class CategoryServiceImplTest {

    @Autowired
    CategoryServiceImpl categoryService;

    @Test
    public void findById() {
        ProductCategory productCategory = categoryService.findById(1);
        System.out.println(productCategory.toString());
        Assert.assertNotNull(productCategory);
    }

    @Test
    public void findByCategoryTypeIn() {
        List<ProductCategory> productCategoryList = categoryService.findByCategoryTypeIn(Arrays.asList(14122,14123));
        Assert.assertNotEquals(0,productCategoryList.size());
    }

    @Test
    public void save() {
        ProductCategory productCategory = new ProductCategory();
        productCategory.setCategoryName("纯茶");
        productCategory.setCategoryType(14124);
        productCategory.setCategoryIcon("www.pic.png");
        ProductCategory result = categoryService.save(productCategory);
        System.out.println(result.toString());
        Assert.assertNotNull(result);
    }

    @Test
    public void findAll() {
        List<ProductCategory> productCategoryList = categoryService.findAll();
        Assert.assertNotEquals(0,productCategoryList.size());
    }

    @Test
    public void deleteById() {
        categoryService.deleteById(12);
    }
}